requirements.txt
Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
stripe==7.6.0
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
sentry-sdk[flask]==1.38.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
